requirements/test.txt
-r prod.txt

django-debug-toolbar==4.2.0
flake8-commas==2.1.0
flake8-dunder-all==0.3.0
flake8-import-order==0.18.2
flake8-print==5.0.0
flake8-quotes==3.3.2
flake8==6.1.0
parameterized==0.9.0
pep8-naming==0.13.3